+++ runtime/runtime_test.go
@@ -25,7 +25,7 @@ func TestExecute(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	rt := runtime.New(state, &block.Header{}, func(uint64) thor.Hash { return thor.Hash{} })
+	rt := runtime.New(state, &block.Header{}, func(uint32) thor.Hash { return thor.Hash{} })
 
 	addr := thor.BytesToAddress([]byte("acc1"))
 	amount := big.NewInt(1000 * 1000 * 1000 * 1000)
@@ -106,7 +106,7 @@ func TestExecuteTransaction(t *testing.T) {
 	sig, _ := signing.Sign(tx, crypto.FromECDSA(key))
 	tx = tx.WithSignature(sig)
 
-	rt := runtime.New(state, &block.Header{}, func(uint64) thor.Hash { return thor.Hash{} })
+	rt := runtime.New(state, &block.Header{}, func(uint32) thor.Hash { return thor.Hash{} })
 	receipt, _, err := rt.ExecuteTransaction(tx, signing)
 	if err != nil {
 		t.Fatal(err)
